+++ CHANGELOG.md
@@ -12,6 +12,7 @@
   * Contact Storage to 1.0-beta8 (from 1.0-beta9)
   * Metatag to 1.1 (from 1.0)
   * Profile to 1.0-beta1 (from 1.0-alpha7)
+* Add a conditional display to the field Icon and Image from paragraph type Icon box.
 * Added an option during the profile installation to choose to install Presto Theme as default theme.
 
 ## 8.x-1.0-alpha2 (1.0.0-alpha.2)